% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{bias_calculator}
\alias{bias_calculator}
\title{bias calculation}
\usage{
bias_calculator(real, predicted)
}
\arguments{
\item{real}{Numeric vector with the "real" values}

\item{predicted}{Numeric vector with the "predicted" values}
}
\description{
Simple function ti calculate the bias value
}
\details{
Usually bias values are calculated between the real values and the predicted
values, but also can be calculated between simple and complex model results
}
