% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_site.R
\name{eval_site}
\alias{eval_site}
\title{Evaluates medfate for one site}
\usage{
eval_site(code, transpMode = "simple", plot = FALSE, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\description{
Evaluates medfate for one site
}
