% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_var.R
\name{eval_var}
\alias{eval_var}
\title{Statistics summary for a given variable}
\usage{
eval_var(x, var, measured_data, trunc = NULL)
}
\arguments{
\item{x}{object of \code{\link{spwb}}}

\item{var}{Character with the variable name}

\item{measured_data}{Data frame with the "real" data}

\item{trunc}{Numeric vector indicating the subset of data to be compared}
}
\description{
Statistics summary for a given variable
}
