% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{global_process}
\alias{global_process}
\title{Main function to do the calibration/validation}
\usage{
global_process(sites, wd, transpMode, SPParams = "old", tapering = TRUE,
  rhizosphere = 0.15)
}
\arguments{
\item{sites}{A character vector with the sites codes}

\item{wd}{Complete path to the validation directory}

\item{transpMode}{Character string indicating the transpiration mode to use}

\item{SPParams}{Character indicating which SpParams table to use. "old" for
the SpParamsMED table from medfate and "new" for the newParams table of
MedfateValidation}

\item{tapering}{Logical indicating if use the taper factor (TRUE) or not
(FALSE)}

\item{rhizosphere}{Number indicating the rhizosphere resistance percentage}
}
\description{
Validate a list of sites/plots
}
