% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_mod_functions.R
\name{inputMod}
\alias{inputMod}
\title{Modification of swb input object}
\usage{
inputMod(swbInput, customParams)
}
\arguments{
\item{swbInput}{An object of class \code{swbInput} as obtained from
\code{\link{forest2swbInput}} or \code{\link{swbInput}} functions from
\code{medfate} package}

\item{customParams}{Data frame with the custom parameters info, usually
the result of \code{\link{load_customParams}}}
}
\description{
This function looks for and modifies input parameters with the values
supplied in the customParams.txt file
}
\seealso{
Other CustomMods: \code{\link{SpParams_mod}}
}
\concept{CustomMods}
