% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{list_sites}
\alias{list_sites}
\title{List the appropriate sites for a given validation}
\usage{
list_sites(validation)
}
\arguments{
\item{validation}{Character indicating for which validation sites are ready}
}
\value{
A character vector with the site names for the given validation, except
  for "all", which returns a tibble with the site ID and the validation as
  columns
}
\description{
This function list the sites that fit for a given validation (global,
transpiration and temperature)
}
