% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_functions.R
\name{load_forest}
\alias{load_forest}
\alias{load_treeData}
\alias{load_shrubData}
\alias{load_seedData}
\alias{load_miscData}
\alias{load_meteoData}
\alias{load_soilData}
\alias{load_soilDataUnilayer}
\alias{load_measuredData}
\alias{load_terrainData}
\alias{load_customParams}
\alias{load_remarks}
\alias{load_list}
\title{Loading data functions}
\usage{
load_forest(site)

load_treeData(site)

load_shrubData(site)

load_seedData(site)

load_miscData(site)

load_meteoData(site)

load_soilData(site)

load_soilDataUnilayer(site)

load_measuredData(site)

load_terrainData(site)

load_customParams(site)

load_remarks(site)

load_list(code, includeTrees = TRUE, includeShrubs = TRUE)
}
\arguments{
\item{site}{Character indicating site code to load.}
}
\description{
This functions search and load the data for the corresponding site
}
\section{Functions}{
\itemize{
\item \code{load_treeData}: 

\item \code{load_shrubData}: 

\item \code{load_seedData}: 

\item \code{load_miscData}: 

\item \code{load_meteoData}: 

\item \code{load_soilData}: 

\item \code{load_soilDataUnilayer}: 

\item \code{load_measuredData}: 

\item \code{load_terrainData}: 

\item \code{load_customParams}: 

\item \code{load_remarks}: 

\item \code{load_list}: 
}}

\concept{Load}
