% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{load_rdatas}
\alias{load_rdatas}
\title{Load all the RData's generated in the validation process}
\usage{
load_rdatas()
}
\description{
This function will be used in the general report with all sites data
}
