% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{MAE_calculator}
\alias{MAE_calculator}
\title{MAE calculation}
\usage{
MAE_calculator(real, predicted)
}
\arguments{
\item{real}{Numeric vector with the "real" values}

\item{predicted}{Numeric vector with the "predicted" values}
}
\description{
Simple function to calculate MAE values
}
\details{
Usually MAE values are calculated comparing the predicted values of one
model with the real values measured. But it can also be calculated to compare
between simple and complex models
}
