% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_functions.R
\name{plot_cohorts_both}
\alias{plot_cohorts_both}
\title{Internal function to plot results for Ecohort comparing both models}
\usage{
plot_cohorts_both(models, measured_data)
}
\arguments{
\item{models}{List with the data frames with the results of the models as
generated by \code{\link{saveRes}}}

\item{measured_data}{Measured data dataframe}

\item{soil}{Soil object used in the model}
}
\description{
To use inside \code{\link{plot_res}} function
}
