% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_functions.R
\name{plot_cohorts_simple}
\alias{plot_cohorts_simple}
\title{Internal function to plot results for Ecohort in the simple model}
\usage{
plot_cohorts_simple(models, measured_data)
}
\arguments{
\item{models}{List with the data frames with the results of the models as
generated by \code{\link{saveRes}}}

\item{measured_data}{Measured data dataframe}
}
\description{
To use inside \code{\link{plot_res}} function
}
