% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggviz_functions.R, R/viz_functions.R
\name{plot_eplanttot_both_gg}
\alias{plot_eplanttot_both_gg}
\alias{plot_eplanttot_both}
\title{Internal function to plot results for Eplanttot comparing both models}
\usage{
plot_eplanttot_both_gg(models, measured_data)

plot_eplanttot_both(models, soil, measured_data)
}
\arguments{
\item{models}{List with the data frames with the results of the models as
generated by \code{\link{saveRes}}}

\item{measured_data}{Measured data dataframe}

\item{soil}{Soil object used in the model}
}
\description{
To use inside \code{\link{plot_res}} function
}
\section{Functions}{
\itemize{
\item \code{plot_eplanttot_both_gg}: 
}}

