% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggviz_functions.R, R/viz_functions.R
\name{plot_eplanttot_complex_gg}
\alias{plot_eplanttot_complex_gg}
\alias{plot_eplanttot_complex}
\title{Internal function to plot results for Eplanttot in the complex model}
\usage{
plot_eplanttot_complex_gg(models, measured_data)

plot_eplanttot_complex(models, measured_data)
}
\arguments{
\item{models}{List with the data frames with the results of the models as
generated by \code{\link{saveRes}}}

\item{measured_data}{Measured data dataframe}
}
\description{
To use inside \code{\link{plot_res}} function
}
\section{Functions}{
\itemize{
\item \code{plot_eplanttot_complex_gg}: 
}}

