% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggviz_functions.R
\name{plot_eplanttot_simple_gg}
\alias{plot_eplanttot_simple_gg}
\alias{plot_eplanttot_complex_gg}
\alias{plot_eplanttot_both_gg}
\title{Internal function to plot results for Eplanttot in the simple model}
\usage{
plot_eplanttot_simple_gg(models, measured_data, leaf_norm)

plot_eplanttot_complex_gg(models, measured_data, leaf_norm)

plot_eplanttot_both_gg(models, measured_data, leaf_norm)
}
\arguments{
\item{models}{List with the data frames with the results of the models as
generated by \code{\link{saveRes}}}

\item{measured_data}{Measured data dataframe}

\item{leaf_norm}{Boolean indicating which E value to represent, raw E (FALSE)
or normalized E to leaf area (TRUE)}
}
\description{
To use inside \code{\link{plot_res}} function
}
\section{Functions}{
\itemize{
\item \code{plot_eplanttot_complex_gg}: 

\item \code{plot_eplanttot_both_gg}: 
}}

