% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggviz_functions.R
\name{plot_eval_gg}
\alias{plot_eval_gg}
\title{Plotting the models results}
\usage{
plot_eval_gg(variable, eval, leaf_norm = FALSE)
}
\arguments{
\item{variable}{Character indicating the variable to visualize}

\item{eval}{Evaluation object returned by function \code{\link{eval_site}}}

\item{leaf_norm}{Boolean indicating if transpiration must be normalized by
leaf area previously to plot it. (Only necessary when plotting Transpiration)}
}
\description{
Function to generate models plots in order to inspect the results in a
visual way
}
