% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggviz_functions.R, R/viz_functions.R
\name{plot_res_gg}
\alias{plot_res_gg}
\alias{plot_res}
\title{Plotting the models results}
\usage{
plot_res_gg(variable, models, soil, measured_data, mode)

plot_res(variable, models, soil, measured_data, mode)
}
\arguments{
\item{variable}{Character indicating the variable to visualize}

\item{models}{List with the data frames with the results of the models as
generated by \code{\link{saveRes}}}

\item{soil}{Soil object used in the model}

\item{measured_data}{Measured data dataframe}

\item{mode}{Transpiration mode indicated in the models ('simple', 'complex' or 'both')}
}
\description{
Function to generate models plots in order to inspect the results in a
visual way
}
\section{Functions}{
\itemize{
\item \code{plot_res_gg}: 
}}

