% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggviz_functions.R
\name{plot_res_gg}
\alias{plot_res_gg}
\title{Plotting the models results}
\usage{
plot_res_gg(variable, models, soil, measured_data, mode,
  leaf_norm = FALSE)
}
\arguments{
\item{variable}{Character indicating the variable to visualize}

\item{models}{List with the data frames with the results of the models as
generated by \code{\link{saveRes}}}

\item{soil}{Soil object used in the model}

\item{measured_data}{Measured data dataframe}

\item{mode}{Transpiration mode indicated in the models ('simple', 'complex' or 'both')}

\item{leaf_norm}{Boolean indicating if transpiration must be normalized by
leaf area previously to plot it. (Only necessary when plotting Transpiration)}
}
\description{
Function to generate models plots in order to inspect the results in a
visual way
}
