% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_functions.R
\name{plot_swc_both}
\alias{plot_swc_both}
\title{Internal function to plot results for SWC comparing both models}
\usage{
plot_swc_both(models, soil, measured_data)
}
\arguments{
\item{models}{List with the data frames with the results of the models as
generated by \code{\link{saveRes}}}

\item{soil}{Soil object used in the model}

\item{measured_data}{Measured data dataframe}
}
\description{
To use inside \code{\link{plot_res}} function
}
