% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggviz_functions.R
\name{plot_swc_layers_gg}
\alias{plot_swc_layers_gg}
\title{Plotting SWC for all layers}
\usage{
plot_swc_layers_gg(models)
}
\arguments{
\item{models}{List with the data frames with the results of the models as
generated by \code{\link{saveRes}}}
}
\description{
This function plots the SWC model results for all layers in one plot
}
