% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggviz_functions.R
\name{plot_temp_complex_gg}
\alias{plot_temp_complex_gg}
\title{Plot canopy temperature measured vs modelled}
\usage{
plot_temp_complex_gg(models, measured_data, meteo_data)
}
\arguments{
\item{models}{}

\item{measured_data}{}

\item{meteo_data}{}
}
\description{
Plot canopy temperature measured vs modelled
}
