% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{r_squared_calculator}
\alias{r_squared_calculator}
\title{r squared calculation}
\usage{
r_squared_calculator(real, predicted)
}
\arguments{
\item{real}{Numeric vector with the "real" values}

\item{predicted}{Numeric vector with the "predicted" values}
}
\description{
Simple function to calculate r squared value
}
\details{
Usually r squared values are calculated for the correlation between real
values and predicted values, but also can be calculated between simple and
complex model results
}
