% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{report_eval}
\alias{report_eval}
\title{Main function to do the calibration/validation}
\usage{
report_eval(sites, wd, transpMode, control = medfate::defaultControl())
}
\arguments{
\item{sites}{A character vector with the sites codes}

\item{wd}{Complete path to the validation directory}

\item{transpMode}{Character string indicating the transpiration mode to use}

\item{control}{List with control parameters}
}
\description{
Validate a list of sites/plots
}
