% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{report_render}
\alias{report_render}
\title{Report generation function}
\usage{
report_render(output_file = NULL, output_dir = NULL, ...)
}
\arguments{
\item{output_file}{file name for the output report}

\item{output_dir}{destination folder}

\item{...}{Rmarkdown parameters}
}
\description{
Function for generating automaitcally the report
}
