% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{report_widget}
\alias{report_widget}
\title{Shiny widget for the Status Rmd}
\usage{
report_widget(data, wd)
}
\arguments{
\item{data}{List with the environments where the report objects are located,
generated with \code{\link{load_rdatas}}.}

\item{wd}{character indicating the working directory}
}
\description{
This function generates the shiny widget to inspect sites individually
}
