% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_mod_functions.R
\name{SpParams_mod}
\alias{SpParams_mod}
\title{Modification of SpParams}
\usage{
SpParams_mod(SpParams, customParams)
}
\arguments{
\item{SpParams}{Data frame with the species parameters, usually obtained
from the \code{\link{medfate}} package (\code{data('SpParamsMED')})}

\item{customParams}{Data frame with the custom parameters info, usually
the result of \code{\link{load_customParams}}}
}
\description{
This function modifies the SpParams table with the values supplied in the
customParams.txt file
}
\seealso{
Other CustomMods: \code{\link{inputMod}}
}
\concept{CustomMods}
