% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{temperature_process}
\alias{temperature_process}
\title{Main function to do the Temperature}
\usage{
temperature_process(sites, wd, transpMode = "complex", SPParams = "new")
}
\arguments{
\item{sites}{A character vector with the sites codes}

\item{wd}{Complete path to the validation directory}

\item{transpMode}{Character string indicating the transpiration mode to use}
}
\description{
Validate a list of sites/plots
}
