% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_by_day_functions.R
\name{transp_day_by_day}
\alias{transp_day_by_day}
\title{Transpiration day by day}
\usage{
transp_day_by_day(input, meteoData, soilData, terrainData, measuredData)
}
\arguments{
\item{input}{swb.input object}

\item{meteoData}{Data frame with the meteorological data}

\item{soilData}{Data frame with the soil info data in unilayer format}

\item{terrainData}{Data frame with the terrain info data}

\item{measuredData}{Data frame with the measured data}
}
\description{
Getting the transpiration calculated day by day using measured SWC values
}
