% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{transpiration_process}
\alias{transpiration_process}
\title{Main function to do the transpiration process}
\usage{
transpiration_process(sites, wd, transpMode, SPParams = "old")
}
\arguments{
\item{sites}{A character vector with the sites codes}

\item{wd}{Complete path to the validation directory}

\item{transpMode}{Character string indicating the transpiration mode to use}
}
\description{
Validate a list of sites/plots
}
