% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_site.R
\name{eval_site}
\alias{eval_site}
\title{Evaluates medfate for one site}
\usage{
eval_site(code, transpMode = "simple",
  control = medfate::defaultControl(), plot = FALSE)
}
\arguments{
\item{code}{String with the code of the site to be processed}

\item{transpMode}{Either 'simple', 'complex' or 'both'}

\item{plot}{A boolean flag indicating wether plots should be produced}
}
\value{
A list with evaluation results
}
\description{
Evaluates medfate for one site
}
