% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{global_process}
\alias{global_process}
\title{Main function to do the calibration/validation}
\usage{
global_process(sites, wd, transpMode, SPParams = "old")
}
\arguments{
\item{sites}{A character vector with the sites codes}

\item{wd}{Complete path to the validation directory}

\item{transpMode}{Character string indicating the transpiration mode to use}
}
\description{
Validate a list of sites/plots
}
