% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_functions.R
\name{load_rdatas}
\alias{load_rdatas}
\title{Load all the RData's generated in the validation process}
\usage{
load_rdatas(type = "Global")
}
\arguments{
\item{type}{Character indicating which data to load ("Global", "Temperature"
and "Definitive")}
}
\description{
This function will be used in the general report with all sites data
}
