% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggviz_functions.R, R/viz_functions.R
\name{plot_res_gg}
\alias{plot_res_gg}
\alias{plot_eval_gg}
\alias{plot_res}
\alias{plot_eval}
\title{Plotting the model evaluation results}
\usage{
plot_res_gg(variable, models, soil, measured_data, mode,
  leaf_norm = FALSE)

plot_eval_gg(variable, eval, leaf_norm = FALSE)

plot_res(variable, models, soil, measured_data, mode)

plot_eval(variable, eval)
}
\arguments{
\item{variable}{A string indicating the variable to visualize:
\itemize{
  \item{\code{SWC} : Soil water content.}
  \item{\code{Eplanttot} : Overall plant transpiration.}
  \item{\code{E_by_Cohort} : Plant transpiration by cohort.}
}}

\item{models}{List with the data frames with the results of the models as
generated by \code{\link{saveRes}}}

\item{soil}{Soil object used in the model}

\item{measured_data}{Measured data dataframe}

\item{mode}{Transpiration mode indicated in the models ('simple', 'complex' or 'both')}

\item{leaf_norm}{Boolean indicating if transpiration must be normalized by
leaf area previously to plot it. (Only necessary when plotting Transpiration)}

\item{eval}{Evaluation object returned by function \code{\link{eval_site}}}
}
\description{
Functions to generate model evaluation plots in order to inspect the results in a
visual way
}
