% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{report_render}
\alias{report_render}
\title{Report generation function}
\usage{
report_render(output_file = NULL, output_dir = NULL, ...)
}
\arguments{
\item{output_file}{file name for the output report}

\item{output_dir}{destination folder}

\item{...}{Rmarkdown parameters}
}
\description{
Function for generating automaitcally the report
}
\details{
The parameters accepted by the rmarkdown template (\code{...} argument) are
the following:
\itemize{
  \item{wd: working directory (i.e. the route to the validation root folder)}
  \item{code: Site/Plot code for which the report must be generated}
  \item{transpMode: Transpiration mode, accepted values are \code{simple},
        \code{complex} and \code{both}}
}
}
