% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_functions.R
\name{saveRes}
\alias{saveRes}
\title{Save results}
\usage{
saveRes(simple_res = NULL, complex_res = NULL, spParams = NULL, site_code)
}
\arguments{
\item{simple_res}{swb object with the simple model results}

\item{complex_res}{swb object with the complex model results}

\item{spParams}{data.frame containing the species parameters used in the model}

\item{site_code}{character with the site/plot code name}
}
\description{
Function to save model results in txt files, one for each model mode
}
