% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_functions.R
\name{saveRes}
\alias{saveRes}
\title{Save results}
\usage{
saveRes(simple_res = NULL, complex_res = NULL, measured_vars = NULL,
  spParams = NULL, site_code, write = TRUE)
}
\arguments{
\item{simple_res}{swb object with the simple model results}

\item{complex_res}{swb object with the complex model results}

\item{measured_vars}{character vector with the names of the variables in
measuredData data frame.}

\item{spParams}{data.frame containing the species parameters used in the model}

\item{site_code}{character with the site/plot code name}

\item{write}{boolean indicating if tables must be written in disk, default to
TRUE}
}
\description{
Function to save model results in txt files, one for each model mode
}
\concept{Save}
