% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{statistics_summary}
\alias{statistics_summary}
\title{Statistics summary for a given variable}
\usage{
statistics_summary(var, models, measured_data, soil)
}
\arguments{
\item{var}{Character with the variable name}

\item{models}{list with the dataframes of models results, typically the
result of \code{\link{saveRes}}}

\item{measured_data}{Data frame with the "real" data}

\item{soil}{Soil object needed to calculate SWC from W}
}
\description{
This function calculates the different statistics for a given variable
accounting for the models performed (simple, complex, both)
}
